backend/requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10

# WSGI Server
gunicorn==21.2.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Task Queue
celery==5.3.4
redis==5.0.1

# AI and ML
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
chromadb==0.4.18
sentence-transformers==2.2.2

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23

# AWS
boto3==1.34.0
botocore==1.34.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
Werkzeug==2.3.7

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0 